#  Speed up cost calculation tools in gdp.py and regional_differentiation.py